requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
appdirs=1.4.4=pyh9f0ad1d_0
atk-1.0=2.38.0=hcb7b3dd_1
atpublic=3.0.1=pyhd8ed1ab_0
aws-c-auth=0.7.16=h0d2f7a6_8
aws-c-cal=0.6.10=h677d54c_2
aws-c-common=0.9.14=h93a5062_0
aws-c-compression=0.2.18=h677d54c_2
aws-c-event-stream=0.4.2=h59ac3ca_6
aws-c-http=0.8.1=hfe5d766_7
aws-c-io=0.14.6=h9ac2cdb_1
aws-c-mqtt=0.10.3=hb8a1441_2
aws-c-s3=0.5.2=h4398043_2
aws-c-sdkutils=0.1.15=h677d54c_2
aws-checksums=0.1.18=h677d54c_2
aws-crt-cpp=0.26.3=h0de420c_2
aws-sdk-cpp=1.11.267=h2fb64bc_3
bidict=0.23.1=pyhd8ed1ab_0
brotli-python=1.1.0=py39hb198ff7_1
bzip2=1.0.8=h93a5062_5
c-ares=1.27.0=h93a5062_0
ca-certificates=2024.2.2=hf0a4a13_0
cairo=1.18.0=hd1e100b_0
certifi=2024.2.2=pyhd8ed1ab_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
expat=2.6.2=hebf3989_0
filelock=3.13.1=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_1
fontconfig=2.14.2=h82840c6_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
freetype=2.12.1=hadb7bae_2
fribidi=1.0.10=h27ca646_0
fsspec=2024.3.1=pyhca7485f_0
gdk-pixbuf=2.42.10=hcea6d13_5
gettext=0.21.1=h0186832_0
gflags=2.2.2=hc88da5d_1004
giflib=5.2.1=h1a8c8d9_3
glog=0.7.0=hc6770e3_0
graphite2=1.3.13=h9f76cd9_1001
graphviz=9.0.0=h3face73_1
greenlet=3.0.3=py39hf3050f2_0
gtk2=2.24.33=h7895bb2_4
gts=0.7.6=he42f4ea_4
harfbuzz=8.3.0=h8f0ba13_0
humanize=4.9.0=pyhd8ed1ab_0
ibis-framework-core=8.0.0=pyhd8ed1ab_1
ibis-sqlite=8.0.0=hd8ed1ab_1
icu=73.2=hc8870d7_0
idna=3.6=pyhd8ed1ab_0
importlib-metadata=7.1.0=pyha770c72_0
importlib-resources=6.4.0=pyhd8ed1ab_0
importlib_resources=6.4.0=pyhd8ed1ab_0
jinja2=3.1.3=pyhd8ed1ab_0
joblib=1.3.2=pyhd8ed1ab_0
krb5=1.21.2=h92f50d5_0
lerc=4.0.0=h9a09cb3_0
libabseil=20240116.1=cxx17_hebf3989_2
libarrow=15.0.2=h8eee870_0_cpu
libarrow-acero=15.0.2=hebf3989_0_cpu
libarrow-dataset=15.0.2=hebf3989_0_cpu
libarrow-flight=15.0.2=h1f98dca_0_cpu
libarrow-flight-sql=15.0.2=hb095944_0_cpu
libarrow-gandiva=15.0.2=h2c81988_0_cpu
libarrow-substrait=15.0.2=h50959cf_0_cpu
libblas=3.9.0=21_osxarm64_openblas
libbrotlicommon=1.1.0=hb547adb_1
libbrotlidec=1.1.0=hb547adb_1
libbrotlienc=1.1.0=hb547adb_1
libcblas=3.9.0=21_osxarm64_openblas
libcrc32c=1.1.2=hbdafb3b_0
libcurl=8.6.0=h2d989ff_0
libcxx=16.0.6=h4653b0c_0
libdeflate=1.19=hb547adb_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h93a5062_2
libevent=2.1.12=h2757513_1
libexpat=2.6.2=hebf3989_0
libffi=3.4.2=h3422bc3_5
libgd=2.3.3=hfdf3952_9
libgfortran=5.0.0=13_2_0_hd922786_3
libgfortran5=13.2.0=hf226fd6_3
libglib=2.80.0=hfc324ee_1
libgoogle-cloud=2.22.0=hbebe991_1
libgoogle-cloud-storage=2.22.0=h8a76758_1
libgrpc=1.62.1=h9c18a4f_0
libiconv=1.17=h0d3ecfb_2
libjpeg-turbo=3.0.0=hb547adb_1
liblapack=3.9.0=21_osxarm64_openblas
libllvm16=16.0.6=haab561b_3
libnghttp2=1.58.0=ha4dd798_1
libopenblas=0.3.26=openmp_h6c19121_0
libparquet=15.0.2=h278d484_0_cpu
libpng=1.6.43=h091b4b1_0
libprotobuf=4.25.3=hbfab5d5_0
libre2-11=2023.09.01=h7b2c953_2
librsvg=2.56.3=h55a2576_1
libsqlite=3.45.2=h091b4b1_0
libssh2=1.11.0=h7a5bd25_0
libthrift=0.19.0=h026a170_1
libtiff=4.6.0=ha8a6c65_2
libutf8proc=2.8.0=h1a8c8d9_0
libwebp=1.3.2=hf30222e_1
libwebp-base=1.3.2=hb547adb_0
libxml2=2.12.6=h0d0cfa8_0
libzlib=1.2.13=h53f4e23_5
llvm-openmp=18.1.2=hcd81f8e_0
lz4-c=1.9.4=hb7217d7_0
markdown-it-py=3.0.0=pyhd8ed1ab_0
markupsafe=2.1.5=py39h17cfd9d_0
mdurl=0.1.2=pyhd8ed1ab_0
multipledispatch=0.6.0=py_0
ncurses=6.4.20240210=h078ce10_0
numpy=1.26.4=py39h7aa2656_0
openssl=3.2.1=h0d3ecfb_1
orc=2.0.0=h3d3088e_0
pandas=2.2.1=py39h47e51b9_0
pango=1.52.1=hb067d4f_0
parsy=2.1=pyhd8ed1ab_0
pcre2=10.43=h26f9a81_0
pins=0.8.4=pyhd8ed1ab_0
pip=24.0=pyhd8ed1ab_0
pixman=0.43.4=hebf3989_0
pyarrow=15.0.2=py39h8cf9275_0_cpu
pyarrow-hotfix=0.6=pyhd8ed1ab_0
pygments=2.17.2=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.9.19=hd7ebdb9_0_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python-graphviz=0.20.3=pyh717bed2_0
python-tzdata=2024.1=pyhd8ed1ab_0
python-xxhash=3.4.1=py39h0f82c59_0
python_abi=3.9=4_cp39
pytz=2024.1=pyhd8ed1ab_0
pyyaml=6.0.1=py39h0f82c59_1
re2=2023.09.01=h4cba328_2
readline=8.2=h92ec313_1
regex=2023.12.25=py39h17cfd9d_0
requests=2.31.0=pyhd8ed1ab_0
rich=13.7.1=pyhd8ed1ab_0
setuptools=69.2.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.10=h17c5cce_0
sqlalchemy=2.0.28=py39h17cfd9d_0
sqlglot=20.11.0=pyhd8ed1ab_0
tk=8.6.13=h5083fa2_1
toolz=0.12.1=pyhd8ed1ab_0
typing-extensions=4.10.0=hd8ed1ab_0
typing_extensions=4.10.0=pyha770c72_0
tzdata=2024a=h0c530f3_0
urllib3=2.2.1=pyhd8ed1ab_0
wheel=0.43.0=pyhd8ed1ab_0
xxhash=0.8.2=hb547adb_0
xz=5.2.6=h57fd34a_0
yaml=0.2.5=h3422bc3_2
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h53f4e23_5
zstd=1.5.5=h4f39d0f_0
